requirements.txt
asgiref==3.6.0
attrs==22.2.0
awsebcli==3.20.5
bcrypt==4.0.1
blessed==1.20.0
boto3==1.26.59
botocore==1.29.81
cached-property==1.5.2
cement==2.8.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.3
cryptography==40.0.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.5
django-crispy-forms==1.14.0
django-filter==22.1
django-import-export==3.0.2
django-s3-storage==0.13.11
django-storages==1.13.2
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
jsonschema==3.2.0
MarkupPy==1.14
numpy==1.24.2
odfpy==1.4.1
openpyxl==3.0.10
pandas==2.0.0
paramiko==3.1.0
pathspec==0.10.1
Pillow==9.4.0
psycopg2==2.9.6
pycparser==2.21
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==5.4.1
requests==2.26.0
s3transfer==0.6.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.3
stripe==5.2.0
tablib==3.3.0
termcolor==1.1.0
texttable==1.6.7
tzdata==2023.3
urllib3==1.26.14
uuid==1.30
wcwidth==0.1.9
websocket-client==0.59.0
xlrd==2.0.1
xlwt==1.3.0
